+++ packages/zowe-explorer-ftp-extension/CHANGELOG.md
@@ -1,5 +1,11 @@
 All notable changes to the "zowe-explorer-ftp-extension" extension will be documented in this file.
 
+## TBD Release
+
+### New features and enhancements
+
+### Bug fixes
+
 ## `3.0.0-next.202311171754`
 
 ## `3.0.0-next.202311171523`