+++ packages/zowe-explorer/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the "vscode-extension-for-zowe" extension will be documented in this file.
 
+## TBD Release
+
+### New features and enhancements
+
+- Added back local storage for Zowe Explorer persistent items
+
+### Bug fixes
+
 ## `3.0.0-next.202311171754`
 
 ## `3.0.0-next.202311171523`